// types/2022/types.ts
export type GoldSponsor = {
  id: number;
  name: string;
  description: string;
  logoImagePath: string;
  url: string;
};

export type SilverSponsor = {
  name: string;
  logoImagePath: string;
  url: string;
};

export type BronzeSponsor = {
  name: string;
  logoImagePath: string;
  url: string;
};